% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simple_adder.R
\name{Simple_adder}
\alias{Simple_adder}
\title{A simple adder code for toy package}
\usage{
Simple_adder(item_1, item_2)
}
\arguments{
\item{item_1}{first item to be added/combined.}

\item{item_2}{second item to be added/combined.}
}
\value{
This function returns adding for two numeric and join for other scenarios
}
\description{
A simple adder code for toy package
}
